launcher/requirements.txt
# Meteor v2 Launcher Dependencies
pyyaml>=6.0
requests>=2.28.0
